backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pillow==11.0.0
torch==2.5.1
transformers==4.46.3
safetensors==0.4.5
numpy==1.26.4
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
google-generativeai==0.8.3
aiofiles==24.1.0

